C3d: initial support for where clauses
The basic idea is really similar to parsing everparse::parameter
attributes. The main difference is that where clauses need to have the
'this' keyword in scope. To do that, we push a new parsing scope and add
a dummy 'this' variable declaration (at void type) into it, and then
parse the clause. After parsing, we drop the scope so it doesn't leak.

Many where clauses can be present in a given type: they are conjoined
into a single one before printing the 3D spec.