RFC: C3d: rework scoping via an internal static vector of variables
The vector just keeps information to create a VarDecl at the time when
we wish to enter the C3d scope. For now, we ony use it in C3dParseExpr
which pushes the C3d scope, parses an expression, and pops the scope.
This fixes the leakage of parameters and field names into the rest of
the program.

Further, note that was actually dangerous since resolution of `sizeof`
happens at this stage. Previously, this struct:

    typedef struct
    [[
      everparse::entrypoint,
      everparse::process,
      everparse::parameter(uint32_t len)
    ]]
    _FOO {
      int arr[sizeof len];
    } FOO;

Was actually outputting:

    typedef struct
    [[
      everparse::entrypoint,
      everparse::process,
      everparse::parameter(uint32_t len)
    ]]
    _FOO {
      int arr[4];
    } FOO;

Which is perfectly valid C, but not what intended: the `len` parameter
should only be visible in everparse:: attributes. Now, we get an unbound
identifier error for `len` in the array size.

Note: this still will not put all struct fields in the scope,
only those with an everparse::constraint attribute. One can use
everparse::constraint(1) as a workaround.

About resetting:
================

We need to clear the scope when switching to process a different struct.
I see two ways to do this:

 1) When *parsing* an everparse::process attribute, we clear the scope.
    Note that doing it at handling time would be too late.

 2) Store a static pointer to the last struct we were handling,
    and detect when a change happened. In that case, clear the scope.

My first try was 1), but then I realized there is not callback into the
plugin for attributes without arguments. I suppose we could add a callback
here, similar to parseAttributePayload, but simpler, to just notify that
it has been parsed. That's another Clang patch of course.

I then attempted 2) but couldn't really make it work. I could not find a
way to get the (partial) declaration that's being parsed. Maybe I missed
it?

So... I went with 1) again after adding an argument to
everparse::process, so I can do the reset on parseAttributePayload. The
argument is totally ignored for now, but could potentially be an on/off
flag. We do have to parse it and register it though, so we see it later
when traversing the AST, which is suboptimal.

See tests/basic0.h too.

Any opinions? I think the extra "heads up" callback might be the way to
go here.